requirements.txt
gTTS==2.5.3
moviepy==1.0.3
rich==13.9.2
pyfiglet==1.0.2
python-dotenv==1.0.1
